+++ R/plot.R
@@ -20,7 +20,8 @@
 #'
 #' @import methods
 #' @importFrom utils head
-#' @importFrom ggplot2 aes_string ggtitle ylim guides guide_legend theme
+#' @importFrom ggplot2 aes ggtitle ylim guides guide_legend theme
+#' @importFrom ggplot2 .data
 #'   element_blank
 #' @importFrom graphics par
 #' @importFrom stats sd density is.empty.model median model.offset
@@ -123,12 +124,12 @@ setMethod("plot", signature(x = "bdpnormal"), function(x, type = NULL, print = T
   ### Posterior Type Plots
   ##############################################################################
 
-  post_typeplot <- ggplot(D, aes_string(x = "x", y = "y")) +
+  post_typeplot <- ggplot(D, aes(x = .data[["x"]], y = .data[["y"]])) +
     geom_line(
-      size = 2,
-      aes_string(
-        colour = "information_sources",
-        lty = "information_sources"
+      linewidth = 2,
+      aes(
+        colour = .data[["information_sources"]],
+        lty = .data[["information_sources"]]
       )
     ) +
     facet_wrap(~group, ncol = 1, scales = "free") +
@@ -145,9 +146,9 @@ setMethod("plot", signature(x = "bdpnormal"), function(x, type = NULL, print = T
 
   densityplot <- ggplot(
     subset(D, information_sources == "Posterior"),
-    aes_string(x = "x", y = "y")
+    aes(x = .data[["x"]], y = .data[["y"]])
   ) +
-    geom_line(size = 2, aes_string(colour = "group")) +
+    geom_line(linewidth = 2, aes(colour = .data[["group"]])) +
     ylab("Density (PDF)") +
     xlab("Values") +
     theme_bw() +
@@ -202,9 +203,9 @@ setMethod("plot", signature(x = "bdpnormal"), function(x, type = NULL, print = T
     )
 
     discountfun_plot <- ggplot() +
-      geom_line(data = D1, aes_string(y = "y", x = "x", color = "group"), size = 1) +
-      geom_vline(data = D2, aes_string(xintercept = "p_hat", color = "group"), lty = 2) +
-      geom_hline(data = D3, aes_string(yintercept = "p_hat", color = "group"), lty = 2)
+      geom_line(data = D1, aes(y = .data[["y"]], x = .data[["x"]], color = .data[["group"]]), linewidth = 1) +
+      geom_vline(data = D2, aes(xintercept = .data[["p_hat"]], color = .data[["group"]]), lty = 2) +
+      geom_hline(data = D3, aes(yintercept = .data[["p_hat"]], color = .data[["group"]]), lty = 2)
   }
 
   if (arm2) {
@@ -242,9 +243,9 @@ setMethod("plot", signature(x = "bdpnormal"), function(x, type = NULL, print = T
       D6 <- data.frame(group = "Control", p_hat = median(posterior_control$alpha_discount))
 
       discountfun_plot <- discountfun_plot +
-        geom_line(data = D4, aes_string(y = "y", x = "x", color = "group"), size = 1) +
-        geom_vline(data = D5, aes_string(xintercept = "p_hat", color = "group"), lty = 2) +
-        geom_hline(data = D6, aes_string(yintercept = "p_hat", color = "group"), lty = 2)
+        geom_line(data = D4, aes(y = .data[["y"]], x = .data[["x"]], color = .data[["group"]]), linewidth = 1) +
+        geom_vline(data = D5, aes(xintercept = .data[["p_hat"]], color = .data[["group"]]), lty = 2) +
+        geom_hline(data = D6, aes(yintercept = .data[["p_hat"]], color = .data[["group"]]), lty = 2)
     }
   }
 
@@ -262,7 +263,7 @@ setMethod("plot", signature(x = "bdpnormal"), function(x, type = NULL, print = T
 
   if (print) {
     if (is.null(type)) {
-      op <- par(ask = TRUE)
+      op <- par(ask = interactive())
       plot(post_typeplot)
       plot(densityplot)
       if (!is.null(discountfun_plot)) {
@@ -326,7 +327,8 @@ setMethod("plot", signature(x = "bdpnormal"), function(x, type = NULL, print = T
 #'
 #' @import methods
 #' @importFrom utils head
-#' @importFrom ggplot2 aes_string ggtitle ylim guides guide_legend theme
+#' @importFrom ggplot2 aes ggtitle ylim guides guide_legend theme
+#' @importFrom ggplot2 .data
 #'   element_blank
 #' @importFrom graphics par
 #' @importFrom stats density is.empty.model median model.offset model.response
@@ -419,8 +421,8 @@ setMethod("plot", signature(x = "bdpbinomial"), function(x, type = NULL, print =
   ### Posterior Type Plots
   ##############################################################################
 
-  post_typeplot <- ggplot(D, aes_string(x = "x", y = "y")) +
-    geom_line(size = 2, aes_string(color = "information_sources", lty = "information_sources")) +
+  post_typeplot <- ggplot(D, aes(x = .data[["x"]], y = .data[["y"]])) +
+    geom_line(linewidth = 2, aes(color = .data[["information_sources"]], lty = .data[["information_sources"]])) +
     theme_bw() +
     facet_wrap(~group, ncol = 1, scales = "free") +
     ylab("Density (PDF)") +
@@ -435,9 +437,9 @@ setMethod("plot", signature(x = "bdpbinomial"), function(x, type = NULL, print =
 
   densityplot <- ggplot(
     subset(D, information_sources == "Posterior"),
-    aes_string(x = "x", y = "y")
+    aes(x = .data[["x"]], y = .data[["y"]])
   ) +
-    geom_line(size = 2, aes_string(color = "group")) +
+    geom_line(linewidth = 2, aes(color = .data[["group"]])) +
     ylab("Density (PDF)") +
     xlab("Values") +
     theme_bw() +
@@ -485,9 +487,9 @@ setMethod("plot", signature(x = "bdpbinomial"), function(x, type = NULL, print =
     D3 <- data.frame(group = "Treatment", p_hat = median(posterior_treatment$alpha_discount))
 
     discountfun_plot <- ggplot() +
-      geom_line(data = D1, aes_string(y = "y", x = "x", color = "group"), size = 1) +
-      geom_vline(data = D2, aes_string(xintercept = "p_hat", color = "group"), lty = 2) +
-      geom_hline(data = D3, aes_string(yintercept = "p_hat", color = "group"), lty = 2)
+      geom_line(data = D1, aes(y = .data[["y"]], x = .data[["x"]], color = .data[["group"]]), linewidth = 1) +
+      geom_vline(data = D2, aes(xintercept = .data[["p_hat"]], color = .data[["group"]]), lty = 2) +
+      geom_hline(data = D3, aes(yintercept = .data[["p_hat"]], color = .data[["group"]]), lty = 2)
   }
 
   if (arm2) {
@@ -525,9 +527,9 @@ setMethod("plot", signature(x = "bdpbinomial"), function(x, type = NULL, print =
       D6 <- data.frame(group = "Control", p_hat = median(posterior_control$alpha_discount))
 
       discountfun_plot <- discountfun_plot +
-        geom_line(data = D4, aes_string(y = "y", x = "x", color = "group"), size = 1) +
-        geom_vline(data = D5, aes_string(xintercept = "p_hat", color = "group"), lty = 2) +
-        geom_hline(data = D6, aes_string(yintercept = "p_hat", color = "group"), lty = 2)
+        geom_line(data = D4, aes(y = .data[["y"]], x = .data[["x"]], color = .data[["group"]]), linewidth = 1) +
+        geom_vline(data = D5, aes(xintercept = .data[["p_hat"]], color = .data[["group"]]), lty = 2) +
+        geom_hline(data = D6, aes(yintercept = .data[["p_hat"]], color = .data[["group"]]), lty = 2)
     }
   }
 
@@ -545,7 +547,7 @@ setMethod("plot", signature(x = "bdpbinomial"), function(x, type = NULL, print =
 
   if (print) {
     if (is.null(type)) {
-      op <- par(ask = TRUE)
+      op <- par(ask = interactive())
       plot(post_typeplot)
       plot(densityplot)
       if (!is.null(discountfun_plot)) {
@@ -608,7 +610,8 @@ setMethod("plot", signature(x = "bdpbinomial"), function(x, type = NULL, print =
 #'
 #' @import methods
 #' @importFrom utils head
-#' @importFrom ggplot2 aes_string ggtitle ylim guides guide_legend theme
+#' @importFrom ggplot2 aes ggtitle ylim guides guide_legend theme
+#' @importFrom ggplot2 .data
 #'   element_blank
 #' @importFrom graphics par
 #' @importFrom stats density is.empty.model median model.offset model.response
@@ -631,7 +634,7 @@ setMethod("plot", signature(x = "bdpsurvival"), function(x, type = NULL, print =
 
   ### Organize data for current treatment
   time_t <- sort(unique(args1$S_t$time))
-  survival_times_posterior_flat <- lapply(time_t, ppexp,
+  survival_times_posterior_flat <- ppexp_times(time_t,
                                           posterior_treatment$posterior_flat_hazard,
                                           cuts = c(0, breaks)
   )
@@ -647,7 +650,7 @@ setMethod("plot", signature(x = "bdpsurvival"), function(x, type = NULL, print =
   ### Organize data for historical treatment
   if (!is.null(args1$S0_t)) {
     time0_t <- sort(unique(args1$S0_t$time))
-    survival_times_prior <- lapply(time0_t, ppexp,
+    survival_times_prior <- ppexp_times(time0_t,
                                    posterior_treatment$prior_hazard,
                                    cuts = c(0, breaks)
     )
@@ -664,7 +667,7 @@ setMethod("plot", signature(x = "bdpsurvival"), function(x, type = NULL, print =
   }
 
   ### Organize data for treatment posterior
-  survival_times_posterior <- lapply(time_t, ppexp, posterior_treatment$posterior_hazard, cuts = c(0, breaks))
+  survival_times_posterior <- ppexp_times(time_t, posterior_treatment$posterior_hazard, cuts = c(0, breaks))
   survival_median_posterior <- 1 - sapply(survival_times_posterior, median)
 
   D3 <- data.frame(
@@ -677,7 +680,7 @@ setMethod("plot", signature(x = "bdpsurvival"), function(x, type = NULL, print =
   ### Organize data for current control
   if (!is.null(args1$S_c)) {
     time_c <- sort(unique(args1$S_c$time))
-    survival_times_posterior_flat <- lapply(time_c, ppexp,
+    survival_times_posterior_flat <- ppexp_times(time_c,
                                             posterior_control$posterior_flat_hazard,
                                             cuts = c(0, breaks)
     )
@@ -696,7 +699,7 @@ setMethod("plot", signature(x = "bdpsurvival"), function(x, type = NULL, print =
   ### Organize data for historical control
   if (!is.null(args1$S0_c)) {
     time0_c <- sort(unique(args1$S0_c$time))
-    survival_times_prior <- lapply(time0_c, ppexp,
+    survival_times_prior <- ppexp_times(time0_c,
                                    posterior_control$prior_hazard,
                                    cuts = c(0, breaks)
     )
@@ -714,7 +717,7 @@ setMethod("plot", signature(x = "bdpsurvival"), function(x, type = NULL, print =
 
   ### Organize data for control posterior
   if (!is.null(args1$S_c) & !is.null(args1$S0_c)) {
-    survival_times_posterior <- lapply(time_c, ppexp, posterior_control$posterior_hazard, cuts = c(0, breaks))
+    survival_times_posterior <- ppexp_times(time_c, posterior_control$posterior_hazard, cuts = c(0, breaks))
     survival_median_posterior <- 1 - sapply(survival_times_posterior, median)
 
     D6 <- data.frame(
@@ -724,7 +727,7 @@ setMethod("plot", signature(x = "bdpsurvival"), function(x, type = NULL, print =
       y = survival_median_posterior
     )
   } else if (is.null(args1$S_c) & !is.null(args1$S0_c)) {
-    survival_times_posterior <- lapply(time0_c, ppexp, posterior_control$posterior_hazard, cuts = c(0, breaks))
+    survival_times_posterior <- ppexp_times(time0_c, posterior_control$posterior_hazard, cuts = c(0, breaks))
     survival_median_posterior <- 1 - sapply(survival_times_posterior, median)
 
     D6 <- data.frame(
@@ -734,7 +737,7 @@ setMethod("plot", signature(x = "bdpsurvival"), function(x, type = NULL, print =
       y = survival_median_posterior
     )
   } else if (!is.null(args1$S_c) & is.null(args1$S0_c)) {
-    survival_times_posterior <- lapply(time_c, ppexp, posterior_control$posterior_hazard, cuts = c(0, breaks))
+    survival_times_posterior <- ppexp_times(time_c, posterior_control$posterior_hazard, cuts = c(0, breaks))
     survival_median_posterior <- 1 - sapply(survival_times_posterior, median)
 
     D6 <- data.frame(
@@ -751,8 +754,8 @@ setMethod("plot", signature(x = "bdpsurvival"), function(x, type = NULL, print =
   D$group <- factor(D$group, levels = c("Treatment", "Control"))
 
   ### Plot survival curve
-  survival_curves <- ggplot(D, aes_string(x = "x", y = "y")) +
-    geom_line(size = 1.4, aes_string(color = "source", lty = "source")) +
+  survival_curves <- ggplot(D, aes(x = .data[["x"]], y = .data[["y"]])) +
+    geom_line(linewidth = 1.4, aes(color = .data[["source"]], lty = .data[["source"]])) +
     facet_wrap(~group, ncol = 1, scales = "free") +
     ylab("Survival probability") +
     xlab("Time") +
@@ -802,9 +805,9 @@ setMethod("plot", signature(x = "bdpsurvival"), function(x, type = NULL, print =
     D3 <- data.frame(group = "Treatment", p_hat = c(median(posterior_treatment$alpha_discount)))
 
     discountfun_plot <- ggplot() +
-      geom_line(data = D1, aes_string(y = "y", x = "x", color = "group"), size = 1) +
-      geom_vline(data = D2, aes_string(xintercept = "p_hat", color = "group"), lty = 2) +
-      geom_hline(data = D3, aes_string(yintercept = "p_hat", color = "group"), lty = 2)
+      geom_line(data = D1, aes(y = .data[["y"]], x = .data[["x"]], color = .data[["group"]]), linewidth = 1) +
+      geom_vline(data = D2, aes(xintercept = .data[["p_hat"]], color = .data[["group"]]), lty = 2) +
+      geom_hline(data = D3, aes(yintercept = .data[["p_hat"]], color = .data[["group"]]), lty = 2)
   }
 
   if (arm2) {
@@ -842,9 +845,9 @@ setMethod("plot", signature(x = "bdpsurvival"), function(x, type = NULL, print =
       D6 <- data.frame(group = "Control", p_hat = c(median(posterior_control$alpha_discount)))
 
       discountfun_plot <- discountfun_plot +
-        geom_line(data = D4, aes_string(y = "y", x = "x", color = "group"), size = 1) +
-        geom_vline(data = D5, aes_string(xintercept = "p_hat", color = "group"), lty = 2) +
-        geom_hline(data = D6, aes_string(yintercept = "p_hat", color = "group"), lty = 2)
+        geom_line(data = D4, aes(y = .data[["y"]], x = .data[["x"]], color = .data[["group"]]), linewidth = 1) +
+        geom_vline(data = D5, aes(xintercept = .data[["p_hat"]], color = .data[["group"]]), lty = 2) +
+        geom_hline(data = D6, aes(yintercept = .data[["p_hat"]], color = .data[["group"]]), lty = 2)
     }
   }
 
@@ -862,7 +865,7 @@ setMethod("plot", signature(x = "bdpsurvival"), function(x, type = NULL, print =
 
   if (print) {
     if (is.null(type)) {
-      op <- par(ask = TRUE)
+      op <- par(ask = interactive())
       plot(survival_curves)
       if (!is.null(args1$S0_t) | (!is.null(args1$S_c) & !is.null(args1$S0_c))) {
         plot(discountfun_plot)